Don't use AccessStatus for the filter tests
