+++ items/src/test/scala/weco/api/items/services/WorkLookupTest.scala
@@ -7,20 +7,9 @@ import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
 import weco.akka.fixtures.Akka
 import weco.api.items.fixtures.ItemsApiGenerators
-import weco.api.stacks.models.{
-  CatalogueAccessMethod,
-  CatalogueAccessStatus,
-  CatalogueWork
-}
-import weco.catalogue.display_model.identifiers.{
-  DisplayIdentifier,
-  DisplayIdentifierType
-}
-import weco.catalogue.display_model.locations.{
-  DisplayAccessCondition,
-  DisplayLocationType,
-  DisplayPhysicalLocation
-}
+import weco.api.stacks.models.{CatalogueAccessMethod, CatalogueWork}
+import weco.catalogue.display_model.identifiers.{DisplayIdentifier, DisplayIdentifierType}
+import weco.catalogue.display_model.locations.{CatalogueAccessStatus, DisplayAccessCondition, DisplayLocationType, DisplayPhysicalLocation}
 import weco.catalogue.display_model.work.DisplayItem
 import weco.catalogue.internal_model.generators.IdentifiersGenerators
 import weco.fixtures.TestWith
