+++ search/src/main/scala/weco/api/search/models/DocumentFilter.scala
@@ -1,7 +1,5 @@
 package weco.api.search.models
 
-import weco.catalogue.internal_model.locations.AccessStatus
-
 import java.time.LocalDate
 
 sealed trait WorkFilter
@@ -44,8 +42,8 @@ case class ItemsFilter(values: Seq[String]) extends WorkFilter
 case class ItemsIdentifiersFilter(values: Seq[String]) extends WorkFilter
 
 case class AccessStatusFilter(
-  includes: List[AccessStatus],
-  excludes: List[AccessStatus]
+  includes: List[String],
+  excludes: List[String]
 ) extends WorkFilter
 
 case class PartOfFilter(id: String) extends WorkFilter
